Add tasty test runner
This is preparation for adding additional tests. Our existing tests are
hedgehog property tests, the next ones aren't going to be. The tasty
wrapper will allow us to run these different types of tests as a single
suite.
